Have strscpy return long long instead of ssize_t
ssize_t is only guaranteed to be capable of storing the
range [-1, SSIZE_MAX], meaning -E2BIG might not fit

+++ cldm/cldm_ntbs.c
@@ -2,7 +2,7 @@
 
 char const *cldm_basename(char const *path);
 
-ssize_t cldm_strscpy(char *restrict dst, char const *restrict src, size_t dstsize) {
+long long cldm_strscpy(char *restrict dst, char const *restrict src, size_t dstsize) {
     size_t const srclen = strlen(src);
     size_t i = dstsize;
     char *d = dst;
